+++ src/Aiolos.class.php
@@ -2,14 +2,14 @@
 include_once("lib.php");
 
 /**
- * Class named after the goddess of wisdom Athene
+ * Class named after the god of winds Aiolos
  *
- * Used for communicating with the WolframAlpha-API
+ * Used for communicating with the Google-APIs
  * 
- * This class handles requests for the WolframAlpha-API 
- * it extends the knowledge base of ARGUS.
+ * This class handles requests for the Google-APIs 
+ * it extends the knowledge base of ARGUS regarding geometry.
  *
- * This class handles requests for WolframAlpha and is connected to Hermes.
+ * This class handles requests for Googles-APIs and is connected to Hermes.
  *
  * @function InputMessage   ->
  * @function CallAPI        ->
@@ -31,7 +31,7 @@ include_once("lib.php");
  * @author  Roman Harms
  * @version V 0.0.1.4
  */
-class Athene
+class Aiolos
 {
 }
-?>
+?>